Improved refraction
Added a virtual Contains(Vector3) method on the Traceable abstract base
class. It is use for probing a short distance further on from an
intersection. We can use that to get at the Traceable and thus the index
of refraction we move into after leaving a refractive object. So far
only implemented for a few Traceable types