Added multi axis SAH partitioning
My original partitioner for the AABBHierarchy split the prims by mid
point in the axis which provided the most equally sized sets. This does
work well and my initial SAH partitioning was unable to beat it since it
always splits along the longest axis which may not be the best choice.
I've now implemented a SAH partitioner that checks each axis in turn and
picks the split based on lowest overall cost. This has reduced render
times considerably. There are some bugs still to fix, it works well for
mostly axis aligned triangles in scenes but some tris seem to be getting
lost. It should be easy to find the error with a test driver which
checks rays in the AABBHierarchy and then double checks with an
exhaustive search.