New XML based file format reader in progress
It compiles and loads several aspects of the scene so far but there are
more xml element parsers to write before it is on par with the old
format. It will have the advantage of having defaults for most values
and being able to specify values in several different ways.