XML Includes
Added an xml include node. You have to jump through hoops to get
XInclude to work in .net so I added a simpler implementation myself. It
can be used anywhere in the scene. I have added a default implementation
of some materials and primitives. This will come into it's own when I
add material inheritance.