Fixed UI overriding of rendering options.
Following the new file format changes the UI was broken in that you
could no longer specify rendering settings different to that in the
file. This is now fixed by storing a reference to the IRenderer created
in LoadScene so that we can swap in it and out of use according to menu
items in the window.