Depth of field camera and the basis for pathtracing renderer.
I want to add a path tracing renderer and an easy way to get there is to
start with DOF enabled in the ray tracing renderer. We now have a Buffer
which re render to. I keeps track of the colour of each pixel and the
variance of the set of samples. We can use that to add futher pixel
samplers.